+++ src/Window/MainWindow/MainWindow.h
@@ -80,7 +80,7 @@ namespace sigrid{
             std::map<sf::Mouse::Button, sigrid::Tool> m_tools;
 
             std::map<sf::Mouse::Button, bool> m_isMouseButtonPressedMap;
-            std::map<sf::Mouse::Button, sf::Vector2i> m_mouseButtonPressedPositionMap;
+            std::map<sf::Mouse::Button, sf::Vector2f> m_mouseButtonPressedPositionMap;
 
             std::unique_ptr<ColorManager> m_tileColorManagerPtr;
             std::unique_ptr<ColorManager> m_arrowColorManagerPtr;
